combine dms_base and association
