+++ romancal/associations/_association.py
@@ -11,6 +11,14 @@ from ._exceptions import AssociationNotValidError
 
 __all__ = ["_Association"]
 
+# DMS file name templates
+_ASN_NAME_TEMPLATE_STAMP = "r{program}-{acid}_{stamp}_{type}_{sequence:03d}_asn"
+_ASN_NAME_TEMPLATE = "r{program}-{acid}_{type}_{sequence:03d}_asn"
+
+_DEGRADED_STATUS_OK = "No known degraded exposures in association."
+
+# Default product name
+PRODUCT_NAME_DEFAULT = "undefined"
 
 # Configure logging
 logger = logging.getLogger(__name__)
@@ -55,21 +63,65 @@ class _Association(MutableMapping):
         version_id=None,
         target=None,
     ):
+        self._asn_name = None
         self.version_id = version_id
         self.target = target
         self.data = {
             "asn_type": "None",
             "asn_rule": self.asn_rule,
-            "version_id": self.version_id,
             "code_version": __version__,
+            "degraded_status": _DEGRADED_STATUS_OK,
+            "program": "noprogram",
+            "version_id": self.version_id,
             "target": self.target,
         }
         self.meta = {}
 
     @property
     def asn_name(self):
-        """Suggest filename for the association"""
-        return "unnamed_association"
+        """The association name
+
+        The name that identifies this association. When dumped,
+        will form the basis for the suggested file name.
+
+        Typically, it is generated based on the current state of
+        the association, but can be overridden.
+        """
+        if self._asn_name:
+            return self._asn_name
+
+        program = self.data["program"]
+        version_id = self.version_id
+        asn_type = self.data["asn_type"]
+        # sequence was a class attribute incremented several times based on test order
+        sequence = 1
+        # acidid was always a3001
+        acidid = "a3001"
+        target = self.target
+
+        if version_id:
+            name = _ASN_NAME_TEMPLATE_STAMP.format(
+                program=program,
+                acid=acidid,
+                stamp=version_id,
+                type=asn_type,
+                sequence=sequence,
+                target=target,
+            )
+        else:
+            name = _ASN_NAME_TEMPLATE.format(
+                program=program,
+                acid=acidid,
+                type=asn_type,
+                sequence=sequence,
+                target=target,
+            )
+        return name.lower()
+
+    @asn_name.setter
+    def asn_name(self, name):
+        """Override calculated association name"""
+        self._asn_name = name
 
     @classmethod
     def _asn_rule(cls):
@@ -208,6 +260,14 @@ class _Association(MutableMapping):
         except AssociationNotValidError:
             return False
         return True
+
+    def new_product(self, product_name=PRODUCT_NAME_DEFAULT):
+        """Start a new product"""
+        product = {"name": product_name, "members": []}
+        try:
+            self.data["products"].append(product)
+        except (AttributeError, KeyError):
+            self.data["products"] = [product]
 
     def _add_items(self, items, **kwargs):
         """Force adding items to the association
